Adds confirmation before deleted entities

+++ core/static/js/caseAdmin.js
@@ -38,6 +38,7 @@ function admin_updateEntity(endpoint, entity) {
     if (resp && resp.success) {
       alert("Updated an entity");
       console.log("Updated an entity. \nNew values:", updates);
+      location.reload();
     } else {
       alert("Failed to update an entity. \nError: " + resp.message);
       console.log("Failed to update an entity. \nError:", resp.message);
@@ -49,26 +50,29 @@ function admin_updateEntity(endpoint, entity) {
 }
 
 function admin_deleteEntity(endpoint, entity) {
-  // ajax the deleted entity id to the server
-  fetch(endpoint + entity, {
-    method: "DELETE",
-    headers: {
-      "Accept": "application/json",
-      "Content-Type": "application/json",
-      "X-CSRFToken": getCookie("csrftoken"),
-    },
-    credentials: "same-origin",
-    body: JSON.stringify({}),
-  }).then(r => r.json()).then(resp => {
-    if (resp && resp.success) {
-      alert("Deleted an entity");
-      console.log("Deleted an entity.");
-    } else {
-      alert("Failed to delete an entity. \nError: " + resp.message);
-      console.log("Failed to delete an entity. \nError:", resp.message);
-    }
-  }).catch(err => {
-    alert("Failed to delete an entity. \nFatal Error: " + err);
-    console.log("Failed to delete an entity. \nFatal Error:", err);
-  });
+  if (confirm("Are you sure you want to delete this entity?")) {
+    // ajax the deleted entity id to the server
+    fetch(endpoint + entity, {
+      method: "DELETE",
+      headers: {
+        "Accept": "application/json",
+        "Content-Type": "application/json",
+        "X-CSRFToken": getCookie("csrftoken"),
+      },
+      credentials: "same-origin",
+      body: JSON.stringify({}),
+    }).then(r => r.json()).then(resp => {
+      if (resp && resp.success) {
+        alert("Deleted an entity");
+        console.log("Deleted an entity.");
+        location.reload();
+      } else {
+        alert("Failed to delete an entity. \nError: " + resp.message);
+        console.log("Failed to delete an entity. \nError:", resp.message);
+      }
+    }).catch(err => {
+      alert("Failed to delete an entity. \nFatal Error: " + err);
+      console.log("Failed to delete an entity. \nFatal Error:", err);
+    });
+  }
 }
